com.unity.xr.arkit@1.0.0-preview.11 # Changelog All notable changes to this package will be documented in this file.
The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).

## [1.0.0-preview.5] - 2018-03-26

### This is the first release of the ARKit package for multi-platform AR.

In this release we are shipping a working iteration of the ARKit package for
Unity's native multi-platform AR support.
Included in the package are static libraries, configuration files, binaries
and project files needed to adapt ARKit to the Unity multi-platform AR API.

## [1.0.0-preview.8] - 2018-05-07

### Added
-Created a Legacy XRInput interface to automate the switch between 2018.1 and 2018.2 XRInput versions.

## [1.0.0-preview.8] - 2018-05-24
### Added
- Availability check to determine runtime support for ARKit.
- Normalize average brightness reading from 0..1

## [1.0.0-preview.9] - 2018-06-06
- Remove extraneous debug log

## [1.0.0-preview.10] - 2018-06-08
- Do not include build postprocessor when not on iOS
- Add support for reference points attached to planes

## [1.0.0-preview.11] - 2018-06-14
- Fail the build if Camera Usage Description is blank